Corrigido bug na instrucao i2d

+++ instruction_set.c
@@ -929,7 +929,7 @@ int i2f(execution *e){
 int i2d(execution *e){
     operand_type op1 = pop_op(&(e->frame->top));
     operand_type op2;
-    op2.Float = (double) op1.Int;
+    op2.Double = (double) op1.Int;
     push_op(&(e->frame->top),op2,2);
 	return 0;
 }
